+++ src/components/Card.jsx
@@ -4,11 +4,10 @@ import closeIcon from "../assets/close.svg";
 import editIcon from "../assets/edit.svg";
 import jsonLocalStorage from "../utils/jsonLocalStorage";
 import { useState } from "react";
+import EditCardForm from "./EditCardForm";
 
 const Card= ({card,cards,setCards,type,boards,setBoards}) => {
     const [isEdit,setIsEdit] = useState(false);
-    const [title, setTitle] = useState(card.title);
-    const [body, setBody] = useState(card.body);
 
     const handleRemoveBtnClick = () => {
       const nextCards = cards.filter(e=>e.cardId !==card.cardId);
@@ -19,31 +18,11 @@ const Card= ({card,cards,setCards,type,boards,setBoards}) => {
       board.cards=nextCards;
       const nextBoards = boards.filter((e)=>e.type!==type);
       setBoards([...nextBoards,board]);
-      console.log(boards);
-    };
-
-    const handleEditSubmit = (e) => {
-      e.preventDefault();
-      let nextCards = cards.filter(e=>e.cardId !==card.cardId);
-      nextCards = nextCards.concat({ cardId: card.cardId, title: title, body: body })
-      setCards(nextCards);
-      jsonLocalStorage.setItem(type, nextCards);
-      setIsEdit(false);
-
-      const board = boards.find((e)=>e.type===type);
-      board.cards=nextCards;
-      const nextBoards = boards.filter((e)=>e.type!==type);
-      setBoards([...nextBoards,board]);
-      console.log(boards);
     };
 
     const handleEditBtnClick = () => {
       setIsEdit(true);
     };
-
-    const handleEditCancelBtnClick = () => {
-      setIsEdit(false);
-    }; 
 
     const handleMoveBtnClick = (end) => {
       //현재 배열에서 삭제
@@ -65,35 +44,11 @@ const Card= ({card,cards,setCards,type,boards,setBoards}) => {
 
       const nextBoards = boards.filter((e)=>e.type!==type && e.type!==end);
       setBoards([...nextBoards,nextStartBoard,nextEndBoard]);
-      console.log(boards);
     }
 
   return <>
   {isEdit?
-    <form className="editCardForm" onSubmit={handleEditSubmit}>
-    <div className="colors">
-    </div>
-    <input
-      className="editInputTitle"
-      placeholder="해야할 일을 입력하세요."
-      name="title"
-      value={title}
-      onChange={(e) => {
-        setTitle(e.target.value);
-      }}
-    />
-    <textarea
-      className="editInputBody"
-      placeholder="자세히 입력해주세요."
-      name="body"
-      value={body}
-      onChange={(e) => {
-        setBody(e.target.value);
-      }}
-    />
-    <button type="submit" >수정완료</button>
-    <button onClick={handleEditCancelBtnClick}>취소</button>
-  </form>
+   <EditCardForm card={card} cards={cards} setCards={setCards} type={type} boards={boards} setBoards={setBoards} setIsEdit={setIsEdit} />
   :
     <div className="card">
       <div className="cardBtns">
@@ -105,6 +60,7 @@ const Card= ({card,cards,setCards,type,boards,setBoards}) => {
         </button>
       </div>
       <h3 className="cardTitle">{card.title}</h3>
+      {/* 맵처리 가능할 듯 */}
       <div
         className="moveTodo"
         onClick={() => handleMoveBtnClick("todo")}
